Switch from json.org library to minimal json
minimal json is faster, and we were on a really old version of json.org.  Rather than upgrade (which would be a similar API change), I prefer to just switch to minimal json (which is what I have been starting to use in other projects as well).

In addition I updated the gradle wrapper as well.